finished chat record mssage storage code

+++ system/blog-entity-parent/blog-entity-message-model/src/main/java/org/cy/micoservice/blog/entity/message/model/provider/resp/ChatRecordPageResp.java
@@ -17,7 +17,7 @@ public class ChatRecordPageResp implements Serializable {
 
   private String content;
 
-  private Integer seqNo;
+  private Long seqNo;
 
   private String avatar;
